[modbus_client] Stamp the device address in play_complex(); type it uint8
Hoisting the templated-address stamp into the play_complex() override makes it
structurally impossible for an action to forget, instead of a convention each
play() must follow; the stamp_address_ helper and its call go away. The
templatable is uint8_t end to end (the schema validates hex_uint8_t), deleting
the narrowing cast.

+++ esphome/components/modbus_client/__init__.py
@@ -52,7 +52,7 @@ async def register_client_action(var, config, args, response_args):
     cg.add(var.set_parent(parent))
     cg.add(
         var.set_target_address(
-            await cg.templatable(config[CONF_ADDRESS], args, cg.uint16)
+            await cg.templatable(config[CONF_ADDRESS], args, cg.uint8)
         )
     )
     if response_conf := config.get(CONF_ON_RESPONSE):
